fix: フレームカウントのバグを修正

+++ frame.c
@@ -23,10 +23,12 @@ void frame_export(Frame* frame) {
 
     struct stat st;
     char* path = "./out";
-    if (stat(path, &st) == 0 && S_ISDIR(st.st_mode)) {
-        system("rm ./out/*");
-    } else {
-        mkdir("./out", 0777);
+    if (frame->frame_count == 0){
+        if (stat(path, &st) == 0 && S_ISDIR(st.st_mode)) {
+            system("rm ./out/*");
+        } else {
+            mkdir("./out", 0777);
+        }
     }
 
     sprintf(file_name, "./out/frame%d.ppm", ++(*frame).frame_count);
